Data API: renaming the namespace to `Controllers` since this is no longer `ResourceManager` specific
