+++ data/Pandora.Api/V1/Controllers/APISchema.cs
@@ -7,7 +7,7 @@ using Pandora.Api.V1.Helpers;
 using Pandora.Data.Models;
 using Pandora.Data.Repositories;
 
-namespace Pandora.Api.V1.ResourceManager;
+namespace Pandora.Api.V1.Controllers;
 
 [ApiController]
 public class ApiSchemaController : ControllerBase
